requirements.txt
certifi==2018.1.18
chardet==3.0.4
cytoolz==0.9.0
eth-abi==0.5.0
eth-keys==0.1.0b4
eth-tester==0.1.0b11
eth-utils==0.7.4
ethereum==2.3.0
future==0.16.0
idna==2.6
lru-dict==1.1.6
pbkdf2==1.3
py-ecc==1.4.2
py-solc==2.1.0
pycryptodome==3.4.9
pyethash==0.1.27
pysha3==1.0.2
PyYAML==3.12
repoze.lru==0.7
requests==2.18.4
rlp==0.6.0
scrypt==0.8.0
semantic-version==2.6.0
toolz==0.9.0
urllib3==1.22
#web3==v4.0.0-beta.7
web3==v3.16.5 
